Move image related classes from ui to core
ui is supposed to be a design library and should not house anything related to image loading.
