+++ core/consumer-rules.pro
@@ -1,2 +1,6 @@
 # Keep the model classes for JSON parsing.
 -keep class * extends com.adyen.checkout.core.common.internal.model.ModelObject  { *; }
+
+-keep public enum com.adyen.checkout.core.internal.image.LogoSize {
+    *;
+}
