+++ core/src/main/java/com/adyen/checkout/core/internal/image/ImageLoader.kt
@@ -6,7 +6,7 @@
  * Created by ozgur on 15/8/2025.
  */
 
-package com.adyen.checkout.ui.internal.image
+package com.adyen.checkout.core.internal.image
 
 import android.graphics.Bitmap
 import androidx.annotation.RestrictTo
