--- ui/consumer-rules.pro
@@ -1,3 +1,0 @@
--keep public enum com.adyen.checkout.ui.internal.image.LogoSize {
-    *;
-}
